Use common AERecord and TRRecord datatype
