+++ src/waveline/__init__.py
@@ -2,5 +2,6 @@
 Top-level module of waveline.
 """
 
+from .datatypes import AERecord, TRRecord
 from .conditionwave import ConditionWave
 from .spotwave import SpotWave
