+++ tests/unit/test_spotwave.py
@@ -251,6 +251,7 @@ def test_get_ae_data(serial_mock):
     # status record
     s = ae_data[0]
     assert s.type_ == "S"
+    assert s.chan == 1
     assert s.time == pytest.approx(2010240 / 2e6)
     assert s.amplitude == pytest.approx(21 * ADC_TO_VOLTS)
     assert s.rise_time == pytest.approx(502689 / 2e6)
@@ -263,6 +264,7 @@ def test_get_ae_data(serial_mock):
     # hit record
     h = ae_data[1]
     assert h.type_ == "H"
+    assert h.chan == 1
     assert h.time == pytest.approx(3044759 / 2e6)
     assert h.amplitude == pytest.approx(3557 * ADC_TO_VOLTS)
     assert h.rise_time == pytest.approx(24 / 2e6)
@@ -291,6 +293,7 @@ def test_get_tr_data(serial_mock, raw):
     tr_data = sw.get_tr_data(raw=raw)
     serial_mock.write.assert_called_with(b"get_tr_data\n")
 
+    assert tr_data[0].chan == 1
     assert tr_data[0].trai == 1
     assert tr_data[0].time == pytest.approx(43686000 / 2e6)
     assert tr_data[0].samples == 13
@@ -300,6 +303,7 @@ def test_get_tr_data(serial_mock, raw):
     else:
         assert_allclose(tr_data[0].data, data[0] * ADC_TO_VOLTS)
 
+    assert tr_data[1].chan == 1
     assert tr_data[1].trai == 2
     assert tr_data[1].time == pytest.approx(43686983 / 2e6)
     assert tr_data[1].samples == 27
